requirements.txt
langchain
langchain_groq
sentence-transformers
faiss-cpu
openai
docx2txt
PyPDF2
scikit-learn
python-dotenv
requests
torch

